Remove no useful test cases and add new testcase
